fix async examples
Some examples are broken because they call fromCallback or
fromNodeCallback functions with too many arguments.

For example:

```javascript
var stat = Rx.Observable.fromNodeCallback(fs.stat);
var files = ['file1.txt', 'file2.txt', 'file3.txt'];

Rx.Observable
  .for(files, stat) // won't work because calls stat(x, i, obs)
```

The `stat` fn is called with too many arguments, it is only
expecting file, but ends up being called with (file, i, obs)
and that causes stat to use the index `i` as the cb and not
use the appended cb which the fromNodeCallback adds.

So correct code must ensure that only the right number
of arguments are passed so the cb will be appended to proper
location.

```javascript
Rx.Observable
  .for(files, function (file) { return stat(file); })
```

Also in some cases the definition of `files` was missing so I
added the definition for clarity.