some() predicate is not optional
Apparently the predicate is not optional for the `some()` operator, as it throws "TypeError: fn must be a function" if you try to omit a predicate with a source that actually produces values. Removing example that shows using `some()` without a predicate on an empty source since that seems pointless - unless I'm missing something - and potentially misleading.

Either this or `some()` should default the predicate to `Rx.helpers.identity`.

Also removing sentence about this operator having an alias called `some`, since this *is* `some`.