Update readme.md
I added link to the book "RxJS in Action" written by Paul Daniels and Luis Atencio and published by Manning Publications. The book is full of theory and practical examples that build on each other and helps readers begin thinking in a reactive manner. I believe it will offer significant support to RxJS users and the community. The book is already available to the public through Manning Early Access Program (MEAP). Please consider adding this link to the Resources/Books section.